tree-wide: move files to same location as in-tree
Make sure to keep the same file-system hierarchy as we would do upstream.
This makes it a lot easier to generate patches which apply to both,
in-tree and out-of-tree.

This breaks the build. A followup commit will make sure everything works
again. They have not been squashed to preserve history and make git detect
the renames properly.

Signed-off-by: David Herrmann <dh.herrmann@gmail.com>