kdbus/selftests: properly reset flags on re-use
If a test runs NAME_ACQUIRE multiple times, we really need to reset the
'flags' argument for each call. It's an in/out argument, so it might have
been changed by the kdbus_name_acquire() helper. We must not rely on the
flags to be unchanged, so reset them properly.

Signed-off-by: David Herrmann <dh.herrmann@gmail.com>