kdbus: split off kdbus_node_deactivate() from draining
Right now, we always perform node deactivation (marking it as
deactivated) and draining (waiting till all active references are
dropped) at the same time. Hence, the kdbus_node_deactivate() call
performed both operations. This is unfortunate, though. The kdbus_conn
object needs both as separate operations as it must hold a look during
activation but not during draining, which is why kdbus_conn didn't use
kdbus_node so far.

To avoid duplicating code in kdbus_conn, this commit prepares kdbus_node
for use with kdbus_conn. This commit splits kdbus_node_deactivate() and
kdbus_node_drain() so they can be used independently. The latter is a
strict superset of the former, though.

However, this is non-trivial to do. So far, a node was unlinked from its
parent once it was deactivated (but not freed, yet). Therefore, if you
iterate a tree while deactivating nodes (but *not* draining them), the
node you deactivated might get destroyed at the same time by another
thread. Thus, if you continue traversing the tree, you cannot find the
next node as the current node is no longer linked.

By merging deactivation and draining, we could be sure that all nodes got
unlinked from a parent before we continue to the next. Hence, we simply
always iterated into the first child as long as there is a child left. By
splitting of deactivation, this is no longer possible. Instead, we need a
different fix for this: Keeping nodes linked at all times.

This patch changes the lifetime behavior of nodes. Instead of unlinking
them from their parent at the time they're drained, we now keep them
linked until they get freed. Therefore, we can properly traverse node
trees as long as we hold a real reference to the current node. But this
has the side-effect that the name of the node stays occupied as long as
someone holds a reference to the node (like the VFS cache). This is
unacceptable. Therefore, we force deactivated nodes to release their name
slot. Hence, we only match name-lookups on a node if the node is *not*
drained, which thusly allows us to link new nodes even though there's an
old drained node with the same name. The rb-tree stays ordered and we
treat deactivated nodes as *smaller* than their name. As we know each
name can only exist once (in activated state), it does not affect the
rb-tree order.

Signed-off-by: David Herrmann <dh.herrmann@gmail.com>