kdbus: make connections use kdbus_node
The kdbus_node implementation provides a generic lifetime manager, which
we used so far for all kdbus API objects, except kdbus_conn. This has
mostly historic reasons, but also because nodes did not support
conditional deactivation until recently. Now that kdbus_node has all the
required support, this makes kdbus_conn use kdbus_node as well.

The reason behind this is to drop all the code duplication that we had
in kdbus_conn so far. We had to re-implement the active-counter and
lifetime management, even though kdbus_node provides a perfect
implementation for this. Furthermore, this is non-trivial code, which we
documented extensively in kdbus_node. However, kdbus_conn always lacked
proper documentation of the functionality as well as a detailed
description of the interactions between kdbus_node and kdbus_conn.

Now that kdbus_conn uses kdbus_node, we can get rid of all that
duplicated code and remain with a single, well-defined lifetime behavior.
Furthermore, the kdbus_node implementation has always been superior to
kdbus_conn as it closes races during deactivation. This has the benefit
that we can also get rid of the ECONNRESET error code, as shutdown is
synchronous now. As such, a method transaction either sees ENXIO (the
peer is already disconnected and couldn't be found), or the message is
properly transmitted. ECONNRESET (which meant the peer was found but
disconnected at exactly this time) is no longer possible. It never was
possible to trigger it reliably, so the tests were faulty, anyway. With
this fixed, we can get rid of it entirely.

Net-change: Removal of >100 lines of code.

Signed-off-by: David Herrmann <dh.herrmann@gmail.com>